# Split up replay table