--- a/project/thscoreboard/replays/templates/replays/replay_table.html
+++ b/project/thscoreboard/replays/templates/replays/replay_table.html
@@ -1,5 +1,5 @@
 {% load humanize %}
-<table id="replay-table" class="replay-table">
+<table class="replay-table">
     <thead>
         <tr>
             <th>
@@ -27,9 +27,10 @@
             </th>
         </tr>
     </thead>
-    <tbody>
+    <tbody id="replay-table-body">
     </tbody>
 </table>
+
 
 <!-- This script is a hack to get variables from python into a js file  -->
 <script>