Currently, the performance of rendering a large table is bad. For Th08, where I have 4656 replays, it takes quite long for a table to render. During this time, the page does not update and the UI is unresponsive.

To mitigate this, a large replay table is split up into many smaller replay tables that are loaded in series. For every 50 rows, a new table is created. This reduces the time for the page to update, and freezes the UI for less long.

| |Before | After | 
|---|---|---|
| Time taken for page to update (pointerup -> first table rendered) | 0.75s | 0.02s |
| Time taken for page to fully render (pointerup -> last table rendered) | 0.75s  | 0.73s |
| Total blocking time | 0.65s  | 0.33s |


Related issue: #309