@jgdhs27 2023-04-03
one thing i forgot to mention is that the table column widths are broken in this pr. it is required that table columns have fixed widths. however, any changes i make to styling will just clash with #306, so I am putting that off for now.