@32th-System 2023-04-08
a really big table might be an issue on phones but on desktop it isn't too bad and you can Ctrl + F more easily