@raviddog 2023-04-08
> a really big table might be an issue on phones but on desktop it isn't too bad and you can Ctrl + F more easily

It's more just I hate when you navigate away from the page and then come back and you have to scroll down 50 times to get back to where you were.